+++ packages/migration/src/transforms/unit-tests/index.ts
@@ -1,7 +1,30 @@
 import ts from 'typescript'
-import {transformTest} from './migrations.js'
+import {transformRootDescribe, transformRuns} from './migrations.js'
+import {isTestDescribeOrContextCall} from './validators.js'
 
-export const migrateTestFile = (sourceFile: ts.SourceFile): ts.SourceFile => {
-  return transformTest(sourceFile)
+const {factory} = ts
+
+export const migrateTestFile = (sourceFile: ts.SourceFile) => {
+  let describeFound = false
+  const updatedStatements = [...sourceFile.statements]
+    .map(statement => {
+      if (isTestDescribeOrContextCall(statement)) {
+        //   update source file with updated beforeEach, then run on everything else in the block
+        describeFound = true
+        return transformRootDescribe(statement)
+      }
+
+      return statement
+    })
+    .map(transformRuns)
+
+  if (!describeFound) {
+    throw new Error('no describe block found')
+  }
+
+  return factory.updateSourceFile(
+    sourceFile,
+    updatedStatements,
+  )
 }
